More config clean up Started to move CLI into its own panel
