+++ modules/wnd_editor.py
@@ -1,7 +1,7 @@
 """----------------------------------------------------------------------------
    wnd_editor.py
 
-   Copyright (C) 2013-2018 Wilhelm Duembeg
+   Copyright (C) 2013-2020 Wilhelm Duembeg
 
    This file is part of gsat. gsat is a cross-platform GCODE debug/step for
    Grbl like GCODE interpreters. With features similar to software debuggers.
@@ -38,372 +38,11 @@ def hex_to_rgb(hex_color):
     return (int(m.group(1), 16), int(m.group(2), 16), int(m.group(3), 16))
 
 
-class gsatStyledTextCtrlSettingsPanel(scrolled.ScrolledPanel):
-    """ Program settings.
+class gsatStcStyledTextCtrl(stc.StyledTextCtrl):
+    """ gsatStcStyledTextCtrl:
+        Text control to display data
     """
 
-    def __init__(self, parent, config_data, key, **args):
-        scrolled.ScrolledPanel.__init__(self, parent,
-                                        style=wx.TAB_TRAVERSAL | wx.NO_BORDER)
-
-        self.configData = config_data
-        self.key = key
-
-        self.InitUI()
-        self.SetAutoLayout(True)
-        self.SetupScrolling()
-
-    def InitUI(self):
-        vBoxSizer = wx.BoxSizer(wx.VERTICAL)
-
-        # Scrolling section
-        text = wx.StaticText(self, label="Scrolling")
-        font = wx.Font(10, wx.DEFAULT, wx.NORMAL, wx.BOLD)
-        text.SetFont(font)
-        vBoxSizer.Add(text, 0, wx.ALL, border=5)
-
-        hBoxSizer = wx.BoxSizer(wx.HORIZONTAL)
-
-        spText = wx.StaticText(self, label="Auto Scroll")
-        hBoxSizer.Add(spText, 0, flag=wx.ALIGN_RIGHT |
-                      wx.ALIGN_CENTER_VERTICAL)
-
-        if self.key == 'code':
-            asList = ["Never", "Always", "On Kill Focus", "On Goto PC"]
-        else:
-            asList = ["Never", "Always", "On Kill Focus"]
-
-        self.asComboBox = wx.ComboBox(self, -1,
-                                      value=asList[self.configData.get(
-                                          '/%s/AutoScroll' % self.key)],
-                                      choices=asList, style=wx.CB_READONLY)
-        hBoxSizer.Add(
-            self.asComboBox, 0,
-            flag=wx.ALL | wx.ALIGN_LEFT | wx.ALIGN_CENTER_VERTICAL, border=5)
-
-        vBoxSizer.Add(hBoxSizer, 0, wx.LEFT | wx.EXPAND |
-                      wx.ALIGN_LEFT, border=20)
-
-        # General Controls
-        text = wx.StaticText(self, label="General")
-        font = wx.Font(10, wx.DEFAULT, wx.NORMAL, wx.BOLD)
-        text.SetFont(font)
-        vBoxSizer.Add(text, 0, wx.ALL, border=5)
-
-        self.fontSelect = wx.FontPickerCtrl(self, size=(300, -1))
-        vBoxSizer.Add(self.fontSelect, 0, wx.LEFT | wx.ALIGN_LEFT,
-                      border=20)
-        font = wx.Font(self.configData.get('/%s/FontSize' % self.key),
-                       wx.DEFAULT, wx.NORMAL, wx.NORMAL, 0,
-                       unicode(self.configData.get('/%s/FontFace' % self.key)))
-
-        font_style_str = self.configData.get('/%s/FontStyle' % self.key)
-
-        if "bold" in font_style_str:
-            font.MakeBold()
-
-        if "italic" in font_style_str:
-            font.MakeItalic()
-
-        self.fontSelect.SetSelectedFont(font)
-
-        vBoxSizer.Add((10, 10), 0, wx.ALL, border=1)
-
-        gBoxSizer = wx.GridSizer(1, 3)
-
-        self.checkReadOnly = wx.CheckBox(self, label="ReadOnly")
-        self.checkReadOnly.SetValue(
-            self.configData.get('/%s/ReadOnly' % self.key))
-        gBoxSizer.Add(self.checkReadOnly, 0, wx.ALIGN_LEFT)
-
-        self.checkLineNumbers = wx.CheckBox(self, label="Line Numbers")
-        self.checkLineNumbers.SetValue(
-            self.configData.get('/%s/LineNumber' % self.key))
-        gBoxSizer.Add(self.checkLineNumbers, 0, wx.ALIGN_LEFT)
-
-        self.checkCaretLine = wx.CheckBox(self, label="Highlight Caret Line")
-        self.checkCaretLine.SetValue(
-            self.configData.get('/%s/CaretLine' % self.key))
-        gBoxSizer.Add(self.checkCaretLine, 0, wx.ALIGN_LEFT)
-
-        vBoxSizer.Add(gBoxSizer, 0, wx.LEFT, border=20)
-
-        # Colors
-        text = wx.StaticText(self, label="Colors")
-        font = wx.Font(10, wx.DEFAULT, wx.NORMAL, wx.BOLD)
-        text.SetFont(font)
-        vBoxSizer.Add(text, 0, wx.ALL, border=5)
-
-        vColorSizer = wx.BoxSizer(wx.VERTICAL)
-        foregroundColorSizer = wx.GridSizer(1, 6, 0, 0)
-        backgroundColorSizer = wx.GridSizer(1, 6, 0, 0)
-        syntaxColorSizer = wx.GridSizer(3, 6, 0, 0)
-
-        # Foreground
-        text = wx.StaticText(self, label="Foreground")
-        vColorSizer.Add(text, 0, flag=wx.ALL, border=5)
-
-        text = wx.StaticText(self, label="Window")
-        foregroundColorSizer.Add(text, 0, flag=wx.ALIGN_CENTER_VERTICAL)
-        self.windowForeground = csel.ColourSelect(
-            self, -1, "",
-            hex_to_rgb(self.configData.get('/%s/WindowForeground' % self.key)))
-
-        foregroundColorSizer.Add(
-            self.windowForeground, 0,
-            flag=wx.ALIGN_CENTER_VERTICAL | wx.ALIGN_LEFT)
-
-        text = wx.StaticText(self, label="Line Numbers")
-        foregroundColorSizer.Add(text, 0, flag=wx.ALIGN_CENTER_VERTICAL)
-        self.lineNumbersForeground = csel.ColourSelect(
-            self, -1, "",
-            hex_to_rgb(self.configData.get(
-                '/%s/LineNumberForeground' % self.key)))
-
-        foregroundColorSizer.Add(
-            self.lineNumbersForeground, 0,
-            flag=wx.ALIGN_CENTER_VERTICAL | wx.ALIGN_LEFT)
-
-        text = wx.StaticText(self, label="Highlight Line")
-
-        foregroundColorSizer.Add(text, 0, flag=wx.ALIGN_CENTER_VERTICAL)
-        self.caretLineForeground = csel.ColourSelect(
-            self, -1, "",
-            hex_to_rgb(self.configData.get(
-                '/%s/CaretLineForeground' % self.key)))
-
-        foregroundColorSizer.Add(
-            self.caretLineForeground, 0,
-            flag=wx.ALIGN_CENTER_VERTICAL | wx.ALIGN_LEFT)
-
-        vColorSizer.Add(
-            foregroundColorSizer, 0, flag=wx.LEFT | wx.EXPAND, border=20)
-
-        # Background
-        text = wx.StaticText(self, label="")
-        vColorSizer.Add(text, 0, flag=wx.ALL, border=5)
-        text = wx.StaticText(self, label="Background")
-        vColorSizer.Add(text, 0, flag=wx.ALL, border=5)
-
-        text = wx.StaticText(self, label="Window")
-        backgroundColorSizer.Add(text, 0, flag=wx.ALIGN_CENTER_VERTICAL)
-        self.windowBackground = csel.ColourSelect(
-            self, -1, "",
-            hex_to_rgb(self.configData.get('/%s/WindowBackground' % self.key)))
-        backgroundColorSizer.Add(
-            self.windowBackground, 0,
-            flag=wx.ALIGN_CENTER_VERTICAL | wx.ALIGN_LEFT)
-
-        text = wx.StaticText(self, label="Line Numbers")
-        backgroundColorSizer.Add(text, 0, flag=wx.ALIGN_CENTER_VERTICAL)
-        self.lineNumbersBackground = csel.ColourSelect(
-            self, -1, "",
-            hex_to_rgb(self.configData.get(
-                '/%s/LineNumberBackground' % self.key)))
-
-        backgroundColorSizer.Add(
-            self.lineNumbersBackground, 0,
-            flag=wx.ALIGN_CENTER_VERTICAL | wx.ALIGN_LEFT)
-
-        text = wx.StaticText(self, label="Highlight Line")
-        backgroundColorSizer.Add(text, 0, flag=wx.ALIGN_CENTER_VERTICAL)
-        self.caretLineBackground = csel.ColourSelect(
-            self, -1, "",
-            hex_to_rgb(self.configData.get(
-                '/%s/CaretLineBackground' % self.key)))
-        backgroundColorSizer.Add(
-            self.caretLineBackground, 0,
-            flag=wx.ALIGN_CENTER_VERTICAL | wx.ALIGN_LEFT)
-
-        vColorSizer.Add(
-            backgroundColorSizer, 0, flag=wx.LEFT | wx.EXPAND, border=20)
-
-        if self.key == 'code':
-            # Syntax highlighting
-            text = wx.StaticText(self, label="")
-            vColorSizer.Add(text, 0, flag=wx.ALL, border=5)
-            text = wx.StaticText(self, label="Syntax highlighting")
-            vColorSizer.Add(text, 0, flag=wx.ALL, border=5)
-
-            text = wx.StaticText(self, label="G Code")
-            syntaxColorSizer.Add(text, 0, flag=wx.ALIGN_CENTER_VERTICAL)
-            self.gCodeHighlight = csel.ColourSelect(
-                self, -1, "",
-                hex_to_rgb(self.configData.get(
-                    '/%s/GCodeHighlight' % self.key))
-            )
-            syntaxColorSizer.Add(self.gCodeHighlight, 0,
-                                 flag=wx.ALIGN_CENTER_VERTICAL | wx.ALIGN_LEFT)
-
-            text = wx.StaticText(self, label="M Code")
-            syntaxColorSizer.Add(text, 0, flag=wx.ALIGN_CENTER_VERTICAL)
-            self.gModeHighlight = csel.ColourSelect(
-                self, -1, "",
-                hex_to_rgb(self.configData.get(
-                    '/%s/MCodeHighlight' % self.key))
-            )
-            syntaxColorSizer.Add(self.gModeHighlight, 0,
-                                 flag=wx.ALIGN_CENTER_VERTICAL | wx.ALIGN_LEFT)
-
-            text = wx.StaticText(self, label="Axis Codes")
-            syntaxColorSizer.Add(text, 0, flag=wx.ALIGN_CENTER_VERTICAL)
-            self.axisHighlight = csel.ColourSelect(
-                self, -1, "", hex_to_rgb(
-                    self.configData.get('/%s/AxisHighlight' % self.key)))
-
-            syntaxColorSizer.Add(
-                self.axisHighlight, 0,
-                flag=wx.ALIGN_CENTER_VERTICAL | wx.ALIGN_LEFT)
-
-            text = wx.StaticText(self, label="Parameters")
-            syntaxColorSizer.Add(text, 0, flag=wx.ALIGN_CENTER_VERTICAL)
-            self.parametersHighlight = csel.ColourSelect(
-                self, -1, "",
-                hex_to_rgb(self.configData.get(
-                    '/%s/ParametersHighlight' % self.key)))
-
-            syntaxColorSizer.Add(self.parametersHighlight, 0,
-                                 flag=wx.ALIGN_CENTER_VERTICAL | wx.ALIGN_LEFT)
-
-            text = wx.StaticText(self, label="Parameters2")
-            syntaxColorSizer.Add(text, 0, flag=wx.ALIGN_CENTER_VERTICAL)
-            self.parameters2Highlight = csel.ColourSelect(
-                self, -1, "",
-                hex_to_rgb(self.configData.get(
-                    '/%s/Parameters2Highlight' % self.key)))
-
-            syntaxColorSizer.Add(self.parameters2Highlight, 0,
-                                 flag=wx.ALIGN_CENTER_VERTICAL | wx.ALIGN_LEFT)
-
-            text = wx.StaticText(self, label="Comments")
-            syntaxColorSizer.Add(text, 0, flag=wx.ALIGN_CENTER_VERTICAL)
-            self.commentsHighlight = csel.ColourSelect(
-                self, -1, "",
-                hex_to_rgb(self.configData.get(
-                    '/%s/CommentsHighlight' % self.key)))
-
-            syntaxColorSizer.Add(self.commentsHighlight, 0,
-                                 flag=wx.ALIGN_CENTER_VERTICAL | wx.ALIGN_LEFT)
-
-            text = wx.StaticText(self, label="G-Code Line #")
-            syntaxColorSizer.Add(text, 0, flag=wx.ALIGN_CENTER_VERTICAL)
-            self.gCodeLineNumberHighlight = csel.ColourSelect(
-                self, -1, "",
-                hex_to_rgb(self.configData.get(
-                    '/%s/GCodeLineNumberHighlight' % self.key)))
-
-            syntaxColorSizer.Add(
-                self.gCodeLineNumberHighlight, 0,
-                flag=wx.ALIGN_CENTER_VERTICAL | wx.ALIGN_LEFT)
-
-            vColorSizer.Add(
-                syntaxColorSizer, 0, flag=wx.LEFT | wx.EXPAND, border=20)
-
-        vBoxSizer.Add(vColorSizer, 0, wx.LEFT | wx.ALIGN_LEFT, border=10)
-
-        # finish up
-        self.SetSizerAndFit(vBoxSizer)
-
-    def UpdateConfigData(self):
-        asValue = self.asComboBox.GetSelection()
-        if asValue > 0:
-            self.configData.set('/%s/AutoScroll' % self.key,
-                                self.asComboBox.GetSelection())
-
-        self.configData.set('/%s/ReadOnly' % self.key,
-                            self.checkReadOnly.GetValue())
-
-        self.configData.set(
-            '/%s/WindowForeground' % self.key,
-            self.windowForeground.GetColour().GetAsString(wx.C2S_HTML_SYNTAX))
-        self.configData.set(
-            '/%s/WindowBackground' % self.key,
-            self.windowBackground.GetColour().GetAsString(wx.C2S_HTML_SYNTAX))
-
-        self.configData.set('/%s/CaretLine' % self.key,
-                            self.checkCaretLine.GetValue())
-        self.configData.set(
-            '/%s/CaretLineForeground' % self.key,
-            self.caretLineForeground.GetColour().GetAsString(
-                wx.C2S_HTML_SYNTAX))
-        self.configData.set(
-            '/%s/CaretLineBackground' % self.key,
-            self.caretLineBackground.GetColour().GetAsString(
-                wx.C2S_HTML_SYNTAX))
-
-        self.configData.set(
-            '/%s/LineNumber' % self.key, self.checkLineNumbers.GetValue())
-        self.configData.set(
-            '/%s/LineNumberForeground' % self.key,
-            self.lineNumbersForeground.GetColour().GetAsString(
-                wx.C2S_HTML_SYNTAX))
-        self.configData.set(
-            '/%s/LineNumberBackground' % self.key,
-            self.lineNumbersBackground.GetColour().GetAsString(
-                wx.C2S_HTML_SYNTAX))
-
-        font = self.fontSelect.GetSelectedFont()
-        font_style_list = []
-        font_style_str = ""
-
-        if font.GetWeight() == wx.BOLD:
-            font_style_list.append("bold")
-
-        if font.GetStyle() == wx.ITALIC:
-            font_style_list.append("italic")
-
-        if len(font_style_list) == 0:
-            font_style_str = "normal"
-        else:
-            font_style_str = ",".join(font_style_list)
-
-        self.configData.set('/%s/FontFace' % self.key, font.GetFaceName())
-        self.configData.set('/%s/FontSize' % self.key, font.GetPointSize())
-        self.configData.set('/%s/FontStyle' % self.key, font_style_str)
-
-        if self.key == 'code':
-            self.configData.set(
-                '/%s/GCodeHighlight' % self.key,
-                self.gCodeHighlight.GetColour().GetAsString(
-                    wx.C2S_HTML_SYNTAX))
-
-            self.configData.set(
-                '/%s/MCodeHighlight' % self.key,
-                self.gModeHighlight.GetColour().GetAsString(
-                    wx.C2S_HTML_SYNTAX))
-
-            self.configData.set(
-                '/%s/AxisHighlight' % self.key,
-                self.axisHighlight.GetColour().GetAsString(wx.C2S_HTML_SYNTAX))
-
-            self.configData.set(
-                '/%s/ParametersHighlight' % self.key,
-                self.parametersHighlight.GetColour().GetAsString(
-                    wx.C2S_HTML_SYNTAX))
-
-            self.configData.set(
-                '/%s/Parameters2Highlight' % self.key,
-                self.parametersHighlight.GetColour().GetAsString(
-                    wx.C2S_HTML_SYNTAX))
-
-            self.configData.set(
-                '/%s/CommentsHighlight' % self.key,
-                self.commentsHighlight.GetColour().GetAsString(
-                    wx.C2S_HTML_SYNTAX))
-
-            self.configData.set(
-                '/%s/GCodeLineNumberHighlight' % self.key,
-                self.gCodeLineNumberHighlight.GetColour().GetAsString(
-                    wx.C2S_HTML_SYNTAX))
-
-
-"""----------------------------------------------------------------------------
-   gsatStcStyledTextCtrl:
-   Text control to display data
-----------------------------------------------------------------------------"""
-
-
-class gsatStcStyledTextCtrl(stc.StyledTextCtrl):
     def __init__(self, parent, config_data, state_data, id=wx.ID_ANY,
                  pos=wx.DefaultPosition, size=wx.DefaultSize, style=0,
                  name=stc.STCNameStr):
