+++ modules/wnd_main.py
@@ -45,10 +45,11 @@ from wx.lib import scrolledpanel as scrolled
 import modules.config as gc
 import modules.machif_config as mi
 import images.icons as ico
+import modules.wnd_main_config as mwc
 import modules.wnd_editor as ed
 import modules.wnd_machine as mc
-import modules.wnd_machine_config as mcc
 import modules.wnd_jogging as jog
+import modules.wnd_cli as cli
 import modules.wnd_compvision as compv
 import modules.machif_progexec as mi_progexec
 
@@ -96,6 +97,7 @@ gID_MENU_SEARCH_TOOLBAR = wx.NewId()
 gID_MENU_RUN_TOOLBAR = wx.NewId()
 gID_MENU_STATUS_TOOLBAR = wx.NewId()
 gID_MENU_OUTPUT_PANEL = wx.NewId()
+gID_MENU_CLI_PANEL = wx.NewId()
 gID_MENU_MACHINE_STATUS_PANEL = wx.NewId()
 gID_MENU_MACHINE_JOGGING_PANEL = wx.NewId()
 gID_MENU_CV2_PANEL = wx.NewId()
@@ -152,236 +154,8 @@ class gsatLog(wx.PyLog):
             self.tc.AppendText(message + '\n')
 
 
-class gsatGeneralSettingsPanel(scrolled.ScrolledPanel):
-    """ General panel settings
-    """
-
-    def __init__(self, parent, configData, **args):
-        scrolled.ScrolledPanel.__init__(
-            self, parent, style=wx.TAB_TRAVERSAL | wx.NO_BORDER)
-
-        self.configData = configData
-
-        self.InitUI()
-        self.SetAutoLayout(True)
-        self.SetupScrolling()
-        # self.FitInside()
-
-    def InitUI(self):
-        font = wx.Font(10, wx.DEFAULT, wx.NORMAL, wx.BOLD)
-        vBoxSizer = wx.BoxSizer(wx.VERTICAL)
-
-        # run time dialog settings
-        st = wx.StaticText(self, label="General")
-        st.SetFont(font)
-        vBoxSizer.Add(st, 0, wx.ALL, border=5)
-
-        # Add file save backup check box
-        self.cbDisplayRunTimeDialog = wx.CheckBox(
-            self, wx.ID_ANY, "Display run time dialog at program end")
-        self.cbDisplayRunTimeDialog.SetValue(
-            self.configData.get('/mainApp/DisplayRunTimeDialog'))
-        vBoxSizer.Add(self.cbDisplayRunTimeDialog, flag=wx.LEFT, border=25)
-
-        # file settings
-        st = wx.StaticText(self, label="Files")
-        st.SetFont(font)
-        vBoxSizer.Add(st, 0, wx.ALL, border=5)
-
-        # Add file save backup check box
-        self.cbBackupFile = wx.CheckBox(
-            self, wx.ID_ANY, "Create a backup copy of file before saving")
-        self.cbBackupFile.SetValue(self.configData.get('/mainApp/BackupFile'))
-        vBoxSizer.Add(self.cbBackupFile, flag=wx.LEFT, border=25)
-
-        # Add file history spin ctrl
-        hBoxSizer = wx.BoxSizer(wx.HORIZONTAL)
-
-        self.scFileHistory = wx.SpinCtrl(self, wx.ID_ANY, "")
-        self.scFileHistory.SetRange(0, 100)
-        self.scFileHistory.SetValue(
-            self.configData.get('/mainApp/FileHistory/FilesMaxHistory'))
-        hBoxSizer.Add(self.scFileHistory, flag=wx.ALL |
-                      wx.ALIGN_CENTER_VERTICAL, border=5)
-
-        st = wx.StaticText(self, wx.ID_ANY, "Recent file history size")
-        hBoxSizer.Add(st, flag=wx.ALL | wx.ALIGN_CENTER_VERTICAL, border=5)
-
-        vBoxSizer.Add(hBoxSizer, 0, flag=wx.LEFT | wx.EXPAND, border=20)
-
-        # tools settings
-        st = wx.StaticText(self, label="Tools")
-        font = wx.Font(10, wx.DEFAULT, wx.NORMAL, wx.BOLD)
-        st.SetFont(font)
-        vBoxSizer.Add(st, 0, wx.ALL, border=5)
-
-        # Add Inch to mm round digits spin ctrl
-        hBoxSizer = wx.BoxSizer(wx.HORIZONTAL)
-        self.scIN2MMRound = wx.SpinCtrl(self, wx.ID_ANY, "")
-        self.scIN2MMRound.SetRange(0, 100)
-        self.scIN2MMRound.SetValue(
-            self.configData.get('/mainApp/RoundInch2mm'))
-        hBoxSizer.Add(self.scIN2MMRound, flag=wx.ALL |
-                      wx.ALIGN_CENTER_VERTICAL, border=5)
-
-        st = wx.StaticText(self, wx.ID_ANY, "Inch to mm round digits")
-        hBoxSizer.Add(st, flag=wx.ALL | wx.ALIGN_CENTER_VERTICAL, border=5)
-
-        vBoxSizer.Add(hBoxSizer, 0, flag=wx.LEFT | wx.EXPAND, border=20)
-
-        # Add mm to Inch round digits spin ctrl
-        hBoxSizer = wx.BoxSizer(wx.HORIZONTAL)
-        self.scMM2INRound = wx.SpinCtrl(self, wx.ID_ANY, "")
-        self.scMM2INRound.SetRange(0, 100)
-        self.scMM2INRound.SetValue(
-            self.configData.get('/mainApp/Roundmm2Inch'))
-        hBoxSizer.Add(self.scMM2INRound, flag=wx.ALL |
-                      wx.ALIGN_CENTER_VERTICAL, border=5)
-
-        st = wx.StaticText(self, wx.ID_ANY, "mm to Inch round digits")
-        hBoxSizer.Add(st, flag=wx.ALL | wx.ALIGN_CENTER_VERTICAL, border=5)
-
-        vBoxSizer.Add(hBoxSizer, 0, flag=wx.LEFT | wx.EXPAND, border=20)
-
-        self.SetSizer(vBoxSizer)
-
-    def UpdateConfigData(self):
-        self.configData.set('/mainApp/DisplayRunTimeDialog',
-                            self.cbDisplayRunTimeDialog.GetValue())
-        self.configData.set('/mainApp/BackupFile',
-                            self.cbBackupFile.GetValue())
-        self.configData.set('/mainApp/FileHistory/FilesMaxHistory',
-                            self.scFileHistory.GetValue())
-        self.configData.set('/mainApp/RoundInch2mm',
-                            self.scIN2MMRound.GetValue())
-        self.configData.set('/mainApp/Roundmm2Inch',
-                            self.scMM2INRound.GetValue())
-
-
-class gsatSettingsDialog(wx.Dialog):
-    """ Dialog to control program settings
-    """
-
-    def __init__(self, parent, configData, id=wx.ID_ANY, title="Settings",
-                 style=wx.DEFAULT_DIALOG_STYLE | wx.RESIZE_BORDER):
-
-        wx.Dialog.__init__(self, parent, id, title, style=style)
-
-        self.configData = configData
-
-        self.InitUI()
-
-    def InitUI(self):
-        sizer = wx.BoxSizer(wx.VERTICAL)
-
-        # init note book
-        self.imageList = wx.ImageList(16, 16)
-        self.imageList.Add(ico.imgGeneralSettings.GetBitmap())
-        # self.imageList.Add(ico.imgPlugConnect.GetBitmap())
-        self.imageList.Add(ico.imgProgram.GetBitmap())
-        self.imageList.Add(ico.imgLog.GetBitmap())
-        self.imageList.Add(ico.imgCli.GetBitmap())
-        self.imageList.Add(ico.imgMachine.GetBitmap())
-        self.imageList.Add(ico.imgMove.GetBitmap())
-        self.imageList.Add(ico.imgEye.GetBitmap())
-
-        # for Windows and OS X, tabbed on the left don't work as well
-        if sys.platform.startswith('linux'):
-            self.noteBook = wx.Notebook(
-                self, size=(700, 400), style=wx.BK_LEFT)
-        else:
-            self.noteBook = wx.Notebook(self, size=(700, 400))
-
-        self.noteBook.AssignImageList(self.imageList)
-
-        # add pages
-        self.AddGeneralPage(0)
-        self.AddProgramPage(1)
-        self.AddOutputPage(2)
-        self.AddCliPage(3)
-        self.AddMachinePage(4)
-        self.AddJoggingPage(5)
-        self.AddCV2Panel(6)
-
-        # self.noteBook.Layout()
-        sizer.Add(self.noteBook, 1, wx.ALL | wx.EXPAND, 5)
-
-        # buttons
-        line = wx.StaticLine(self, -1, size=(20, -1), style=wx.LI_HORIZONTAL)
-        sizer.Add(line, 0, wx.GROW | wx.ALIGN_CENTER_VERTICAL |
-                  wx.LEFT | wx.RIGHT | wx.TOP, border=5)
-
-        btnsizer = wx.StdDialogButtonSizer()
-
-        btn = wx.Button(self, wx.ID_OK)
-        btnsizer.AddButton(btn)
-
-        btn = wx.Button(self, wx.ID_CANCEL)
-        btnsizer.AddButton(btn)
-
-        btnsizer.Realize()
-
-        sizer.Add(btnsizer, 0, wx.ALIGN_CENTER_VERTICAL |
-                  wx.ALIGN_RIGHT | wx.ALL, 5)
-
-        self.SetSizerAndFit(sizer)
-        # self.SetAutoLayout(True)
-        self.Layout()
-
-    def AddGeneralPage(self, page):
-        self.generalPage = gsatGeneralSettingsPanel(
-            self.noteBook, self.configData)
-        self.noteBook.AddPage(self.generalPage, "General")
-        self.noteBook.SetPageImage(page, page)
-
-    def AddProgramPage(self, page):
-        self.programPage = ed.gsatStyledTextCtrlSettingsPanel(
-            self.noteBook, self.configData, "code")
-        self.noteBook.AddPage(self.programPage, "Program")
-        self.noteBook.SetPageImage(page, page)
-
-    def AddOutputPage(self, page):
-        self.outputPage = ed.gsatStyledTextCtrlSettingsPanel(
-            self.noteBook, self.configData, "output")
-        self.noteBook.AddPage(self.outputPage, "Output")
-        self.noteBook.SetPageImage(page, page)
-
-    def AddCliPage(self, page):
-        self.cliPage = jog.gsatCliSettingsPanel(self.noteBook, self.configData)
-        self.noteBook.AddPage(self.cliPage, "Cli")
-        self.noteBook.SetPageImage(page, page)
-
-    def AddMachinePage(self, page):
-        self.machinePage = mcc.gsatMachineSettingsPanel(
-            self.noteBook, self.configData)
-        self.noteBook.AddPage(self.machinePage, "Machine")
-        self.noteBook.SetPageImage(page, page)
-
-    def AddJoggingPage(self, page):
-        self.jogPage = jog.gsatJoggingSettingsPanel(
-            self.noteBook, self.configData)
-        self.noteBook.AddPage(self.jogPage, "Jogging")
-        self.noteBook.SetPageImage(page, page)
-
-    def AddCV2Panel(self, page):
-        self.CV2Page = compv.gsatCV2SettingsPanel(
-            self.noteBook, self.configData)
-        self.noteBook.AddPage(self.CV2Page, " OpenCV2")
-        self.noteBook.SetPageImage(page, page)
-
-    def UpdateConfigData(self):
-        self.generalPage.UpdateConfigData()
-        self.programPage.UpdateConfigData()
-        self.outputPage.UpdateConfigData()
-        self.cliPage.UpdateConfigData()
-        self.machinePage.UpdateConfigData()
-        self.jogPage.UpdateConfigData()
-        self.CV2Page.UpdateConfigData()
-
-
 class gsatMainWindow(wx.Frame, gc.EventQueueIf):
-    """ Main Window Inits the UI and other panels, it also controls the worker
-    threads and resources such as serial port.
+    """ Main Window Inits the UI and other panels.
     """
 
     def __init__(self, parent, wnd_id=wx.ID_ANY, title="",
@@ -490,8 +264,8 @@ class gsatMainWindow(wx.Frame, gc.EventQueueIf):
             self, self.configData, self.stateData, self.cmdLineOptions)
         self.machineJoggingPanel = jog.gsatJoggingPanel(
             self, self.configData, self.stateData)
-        self.Bind(wx.EVT_TEXT_ENTER, self.OnCliEnter,
-                  self.machineJoggingPanel.cliComboBox)
+        # self.cliPanel = cli.gsatCliPanel(
+        #    self, self.configData, self.stateData)
 
         # output Window
         self.outputText = ed.gsatStcStyledTextCtrl(
@@ -513,9 +287,15 @@ class gsatMainWindow(wx.Frame, gc.EventQueueIf):
             .Caption("G-Code").CloseButton(True).MaximizeButton(True)
             .BestSize(600, 600))
 
+        # self.aui_mgr.AddPane(
+        #     self.cliPanel,
+        #     aui.AuiPaneInfo().Name("CLI_PANEL").Bottom().Row(1)
+        #     .Caption("CLI").CloseButton(True).MaximizeButton(True)
+        #     .BestSize(600, 40))
+
         self.aui_mgr.AddPane(
             self.outputText,
-            aui.AuiPaneInfo().Name("OUTPUT_PANEL").Bottom().Position(1)
+            aui.AuiPaneInfo().Name("OUTPUT_PANEL").Bottom().Row(1)
             .Caption("Output").CloseButton(True).MaximizeButton(True)
             .BestSize(600, 200))
 
@@ -623,6 +403,7 @@ class gsatMainWindow(wx.Frame, gc.EventQueueIf):
         viewMenu.AppendCheckItem(gID_MENU_STATUS_TOOLBAR, "Status &Tool Bar")
         viewMenu.AppendSeparator()
         viewMenu.AppendCheckItem(gID_MENU_OUTPUT_PANEL, "&Output")
+        # viewMenu.AppendCheckItem(gID_MENU_CLI_PANEL, "&CLI")
         viewMenu.AppendCheckItem(gID_MENU_MACHINE_STATUS_PANEL,
                                  "Machine &Status")
         viewMenu.AppendCheckItem(gID_MENU_MACHINE_JOGGING_PANEL,
@@ -790,6 +571,8 @@ class gsatMainWindow(wx.Frame, gc.EventQueueIf):
                   id=gID_MENU_STATUS_TOOLBAR)
         self.Bind(wx.EVT_MENU, self.OnOutput,
                   id=gID_MENU_OUTPUT_PANEL)
+        # self.Bind(wx.EVT_MENU, self.OnCli,
+        #           id=gID_MENU_CLI_PANEL)
         self.Bind(wx.EVT_MENU, self.OnMachineStatus,
                   id=gID_MENU_MACHINE_STATUS_PANEL)
         self.Bind(wx.EVT_MENU, self.OnMachineJogging,
@@ -813,6 +596,8 @@ class gsatMainWindow(wx.Frame, gc.EventQueueIf):
                   id=gID_MENU_STATUS_TOOLBAR)
         self.Bind(wx.EVT_UPDATE_UI, self.OnOutputUpdate,
                   id=gID_MENU_OUTPUT_PANEL)
+        # self.Bind(wx.EVT_UPDATE_UI, self.OnCliUpdate,
+        #           id=gID_MENU_CLI_PANEL)
         self.Bind(wx.EVT_UPDATE_UI, self.OnMachineStatusUpdate,
                   id=gID_MENU_MACHINE_STATUS_PANEL)
         self.Bind(wx.EVT_UPDATE_UI, self.OnMachineJoggingUpdate,
@@ -1181,6 +966,7 @@ class gsatMainWindow(wx.Frame, gc.EventQueueIf):
 
     def UpdateUI(self):
         self.gcText.UpdateUI(self.stateData)
+        # self.cliPanel.UpdateUI(self.stateData)
         self.machineStatusPanel.UpdateUI(self.stateData)
         self.machineJoggingPanel.UpdateUI(self.stateData)
         self.CV2Panel.UpdateUI(self.stateData)
@@ -1481,6 +1267,12 @@ class gsatMainWindow(wx.Frame, gc.EventQueueIf):
     def OnOutputUpdate(self, e):
         self.OnViewMenuUpdate(e, self.outputText)
 
+    # def OnCli(self, e):
+    #     self.OnViewMenu(e, self.cliPanel)
+
+    # def OnCliUpdate(self, e):
+    #     self.OnViewMenuUpdate(e, self.cliPanel)
+
     def OnMachineStatus(self, e):
         self.OnViewMenu(e, self.machineStatusPanel)
 
@@ -1512,7 +1304,7 @@ class gsatMainWindow(wx.Frame, gc.EventQueueIf):
 
     def OnSettings(self, e):
         # do settings dialog
-        dlg = gsatSettingsDialog(self, self.configData)
+        dlg = mwc.gsatSettingsDialog(self, self.configData)
 
         result = dlg.ShowModal()
 
@@ -1525,6 +1317,7 @@ class gsatMainWindow(wx.Frame, gc.EventQueueIf):
 
             self.gcText.UpdateSettings(self.configData)
             self.outputText.UpdateSettings(self.configData)
+            # self.cliPanel(self.configData)
             self.machineStatusPanel.UpdateSettings(self.configData)
             self.machineJoggingPanel.UpdateSettings(self.configData)
             self.CV2Panel.UpdateSettings(self.configData)
@@ -2018,13 +1811,6 @@ class gsatMainWindow(wx.Frame, gc.EventQueueIf):
     # -------------------------------------------------------------------------
     # Other UI Handlers
     # -------------------------------------------------------------------------
-    def OnCliEnter(self, e):
-        cliCommand = self.machineJoggingPanel.GetCliCommand()
-
-        if len(cliCommand) > 0:
-            serialData = "%s\n" % (cliCommand)
-            self.SerialWrite(serialData)
-
     def OnClose(self, e):
         if self.machifProgExec is not None:
             self.machifProgExec.eventPut(gc.EV_CMD_EXIT)
@@ -2034,6 +1820,7 @@ class gsatMainWindow(wx.Frame, gc.EventQueueIf):
         if self.stateData.serialPortIsOpen:
             self.SerialClose()
 
+        # self.cliPanel.SaveCli()
         self.machineJoggingPanel.SaveCli()
         self.configData.save()
         self.aui_mgr.UnInit()
